Renaming refactor: static -> resources services -> endpoints
