+++ src/v0/mod.rs
@@ -1,5 +1,5 @@
 pub mod context;
-pub mod services;
+pub mod endpoints;
 
 use warp::filters::BoxedFilter;
 use warp::{Filter, Rejection, Reply};
@@ -8,5 +8,5 @@ pub fn filter(
     ctx: context::SharedContext,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     let root = warp::path!("v0" / ..);
-    services::filter(root.boxed(), ctx)
+    endpoints::filter(root.boxed(), ctx)
 }
